requirements.txt
alembic
blinker
certifi
charset-normalizer
click
Flask
Flask-Cors
flask-marshmallow
Flask-Migrate
Flask-SQLAlchemy
google-api-core
google-auth
google-cloud-core
google-cloud-pubsub
google-cloud-storage
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
gunicorn
idna
itsdangerous
Jinja2
jmespath
Mako
MarkupSafe
marshmallow
packaging
proto-plus
protobuf
psycopg2-binary
PyJWT
PyMuPDF
python-dateutil
requests
six
SQLAlchemy
typing_extensions
urllib3
watchdog
Werkzeug